append to old results, do not delete
