+++ src/services/mediasearch.ts
@@ -13,3 +13,21 @@ export type SearchResult = {
 };
 
 export type ScrapeSearchResult = Pick<SearchResult, 'title' | 'fileSize' | 'hash'>;
+
+export const flattenAndRemoveDuplicates = (arr: ScrapeSearchResult[][]): ScrapeSearchResult[] => {
+	const flattened = arr.reduce((acc, val) => acc.concat(val), []);
+	const unique = new Map<string, ScrapeSearchResult>();
+	flattened.forEach((item) => {
+		if (!unique.has(item.hash)) {
+			unique.set(item.hash, item);
+		}
+	});
+	return Array.from(unique.values());
+};
+
+export const sortByFileSize = (results: ScrapeSearchResult[]): ScrapeSearchResult[] => {
+	results.sort((a, b) => {
+		return b.fileSize - a.fileSize;
+	});
+	return results;
+};
